tests: split lock-transition suite into dedicated file
