+++ tests/test_locking.h
@@ -1,0 +1,8 @@
+#ifndef TEST_LOCKING_H
+#define TEST_LOCKING_H
+
+void test_lock_transitions(void);
+void test_lock_transitions_stub(void);
+void test_lock_transitions_fork(void);
+
+#endif
